+++ changelog.md
@@ -1,7 +1,9 @@
 
 
 # 0.x
-Increased message count from 100 to 100,000
+- Added a Customers stream, fake customer data now streams into topic `customers`
+- Pizzas stream into `pizza-orders`
+- Increased message count from 100 to 1000
 
 # 0.1
 Initial release. 21 Dec 2023.